Slave should include the table name in its status while  processing RBR events.
Author: Davi Arnaut <davi@twitter.com>

From d9895b4410f744a231f789ec9d894dbe633e3b14 Mon Sep 17 00:00:00 2001
From: Davi Arnaut <davi@twitter.com>
Date: Mon, 14 Jan 2013 14:14:02 -0800
Subject: [PATCH] Slave should include the table name in its status while
 processing RBR events.

When processing a row-based event, the SQL Thread state is updated to
reflect what type of event is being processed, but short of dumping the
event from the logs, there is currently no way to identify to which
table the event is being applied to.

In order to allow for easier identification of which table an event is
being applied to, this change extends the SQL Thread state to also
include the fully qualified name of the table associated with an event
being applied.